+++ src/structure_refinement/surfel.h
@@ -33,21 +33,21 @@ class Surfel : public srrg2_core::MessageSinkBase {
   Surfel();
   virtual ~Surfel();
   bool putMessage(srrg2_core::BaseSensorMessagePtr msg) override;
-  void addObservation(Eigen::Isometry3d&, unsigned int, Eigen::Matrix<double, 3, 5>&);  // Add new observation
+  void addObservation(Eigen::Isometry3f&, unsigned int, Eigen::Matrix<float, 3, 5>&);  // Add new observation
   bool checkIfsurfelIdExists(unsigned int, unsigned int);                               // Checks whether such leafId was already associated
   bool checkIfPoseExists(unsigned int);
   float getLargestRadius();
 
   unsigned int id_;
-  Eigen::Vector3d estPosition_;                                    // Initial estimate of position
-  Eigen::Vector3d estNormal_;                                      // Initial estimate of normal
+  Eigen::Vector3f estPosition_;                                    // Initial estimate of position
+  Eigen::Vector3f estNormal_;                                      // Initial estimate of normal
   std::map<unsigned int, std::set<unsigned int>> poseSurfelsIds_;  // Corresponding id's of surfels in given poses | PoseId -> SurfelId its (id in kdTree vector)
                                                                    // It is used for matching other leafs from other kdTrees
-  std::vector<Eigen::Isometry3d> odomPoses_;                       // Odometry poses from which the surfel was observed | Indices correspond to observations_
+  std::vector<Eigen::Isometry3f> odomPoses_;                       // Odometry poses from which the surfel was observed | Indices correspond to observations_
   std::vector<uint> odomPosesIds_;                                 // IDs of odometry poses corresponding to this->odomPoses_  | Used to detect if given surfel has two observations from one pose | Indices correspond to poses_ and observations_
-  std::vector<Eigen::Matrix<double, 3, 5>> observations_;          // Observations from each pose: surfel eigen vectors (3 cols * 3 rows = 9 el), surfel mean (1 col = 3 el), surfel bbox (1 col = 3 el)
+  std::vector<Eigen::Matrix<float, 3, 5>> observations_;          // Observations from each pose: surfel eigen vectors (3 cols * 3 rows = 9 el), surfel mean (1 col = 3 el), surfel bbox (1 col = 3 el)
                                                                    // It might be enough to store only normal (first column of eigen vectors)
-                                                                   // std::vector<Eigen::Vector3d> points_;
+                                                                   // std::vector<Eigen::Vector3f> points_;
 };
 using SurfelPtr = std::shared_ptr<Surfel>;
 
@@ -58,15 +58,15 @@ class SynthSurfel {
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
   SynthSurfel(){};
   ~SynthSurfel(){};
-  void addObservation(Eigen::Matrix<double, 3, 2>&);
-  Eigen::Matrix3d matrixBetween2Vectors(Eigen::Vector3d, Eigen::Vector3d);
+  void addObservation(Eigen::Matrix<float, 3, 2>&);
+  Eigen::Matrix3f matrixBetween2Vectors(Eigen::Vector3f, Eigen::Vector3f);
 
   Eigen::Isometry3f getIsometry();
 
   unsigned int id_;
-  Eigen::Vector3d estPosition_;                            // Initial estimate of position
-  Eigen::Vector3d estNormal_;                              // Initial estimate of normal
-  std::vector<Eigen::Matrix<double, 3, 2>> observations_;  // Normal and mean
+  Eigen::Vector3f estPosition_;                            // Initial estimate of position
+  Eigen::Vector3f estNormal_;                              // Initial estimate of normal
+  std::vector<Eigen::Matrix<float, 3, 2>> observations_;  // Normal and mean
 };
 using SynthSurfelPtr = std::shared_ptr<SynthSurfel>;
