+++ src/structure_refinement/surfelv2.h
@@ -10,7 +10,7 @@
 #include "kdtree.cuh"
 
 namespace structure_refinement {
-using ContainerType = std::vector<Eigen::Vector3d>;
+using ContainerType = std::vector<Eigen::Vector3f>;
 using TreeNodeType = TreeNode3D<ContainerType>;
 using TreeNodeTypePtr = TreeNodeType *;
