+++ src/structure_refinement/utils.h
@@ -4,7 +4,7 @@
 
 namespace structure_refinement {
 
-Eigen::Matrix3d matrixBetween2Vectors(Eigen::Vector3d a, Eigen::Vector3d b);
-double angleBetween2Vectors(const Eigen::Vector3d& a, const Eigen::Vector3d& b);
+Eigen::Matrix3f matrixBetween2Vectors(Eigen::Vector3f a, Eigen::Vector3f b);
+float angleBetween2Vectors(const Eigen::Vector3f& a, const Eigen::Vector3f& b);
 
 }
